requirements.txt
discord
mariadb
datetime
asyncio
pycryptodome